Fix push

+++ database/migrations/2025_06_12_032448_create_products_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->decimal('price',10,2);
             $table->integer('stock')->default(0);
             $table->string('image')->nullable();
-            $table->decimal('rating', 3, 2)->default(0); 
+            $table->float('rating')->default(0); 
             $table->string('color')->nullable();
             $table->timestamps();
